Refactor into client and server packages
