+++ src/mirrulations_client/client.py
@@ -1,6 +1,6 @@
-import mirrulations.document_processor as doc
-import mirrulations.documents_processor as docs
-import mirrulations.api_call_management as man
+import mirrulations_client.document_processor as doc
+import mirrulations_client.documents_processor as docs
+import mirrulations_core.api_call_management as man
 import requests
 import json
 import time
