+++ src/mirrulations_client/document_processor.py
@@ -1,5 +1,5 @@
 import tempfile
-from mirrulations.documents_processor import *
+from mirrulations_client.documents_processor import *
 import mirrulations_core.config as config
 
 base_url = 'https://api.data.gov/regulations/v3/document?documentId='
