+++ src/mirrulations_server/docs_work_gen.py
@@ -2,8 +2,8 @@ import requests
 import random
 import string
 import redis
-import mirrulations.redis_manager as redis_manager
-import mirrulations.endpoints as endpoints
+import mirrulations_server.redis_manager as redis_manager
+import mirrulations_server.endpoints as endpoints
 import mirrulations_core.config as config
 import logging
