+++ tests/mirrulations_tests/test_doc_filter.py
@@ -2,7 +2,7 @@ import pytest
 import requests_mock
 import tempfile
 import os
-import mirrulations.doc_filter as df
+import mirrulations_server.doc_filter as df
 
 
 PATH = os.path.join(os.path.abspath(os.path.dirname(__file__)), "../../tests/test_files/mirrulations_files/")
