+++ tests/mirrulations_tests/test_docs_filter.py
@@ -1,5 +1,5 @@
 import json
-import mirrulations.docs_filter as dsf
+import mirrulations_server.docs_filter as dsf
 import os
 
 PATH = os.path.join(os.path.abspath(os.path.dirname(__file__)), "../../tests/test_files/mirrulations_files/")
